Move conftest for --vfs option into tiledb/tests/conftest.py
This should allow tests to run correctly when installed (currently
failing because the top-level conftest is not applied, so the
`pytest.tiledb_vfs` add-on attribute is not available).

Now requires running `pytest tiledb --vfs=...` from the top-level
directory (`pytest ...` still works w/out `--vfs` option).

- Remove base dir conftest.py
